+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 3.0.4 (unreleased)
 
+- **dashboards**: PCP Vector Host Overview: add pmproxy URL and hostspec variables
+- **vector, bpftrace**: use `pcp://127.0.0.1` as default hostspec (no functional change)
 - **chore**: update dependencies
 - **test**: replace convey with testify for the Go tests